put datasets all back into 1 file After some merge hell, I couldn't keep the seperated files for every dataset and make everything work again. It's better to make a seperate PR at some point focused solely on disentangling these classes again.
